[Feat] #82. v2-draft

+++ backend/orury/src/main/java/com/kernel360/orury/domain/user/db/RefreshTokenRepository.java
@@ -5,5 +5,5 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RefreshTokenRepository extends JpaRepository<RefreshTokenEntity, Long> {
-	Optional<RefreshTokenEntity> findByUserIdAndReissueCountLessThan(Long id, int count);
+	Optional<RefreshTokenEntity> findByUserIdAndReissueCountLessThan(Long id, long count);
 }
